+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Internal changes to use proper-cased views everywhere instead of legacy task=viewname.taskname when building public URLs
 * Do not go through browser fingerprinting if the Remember Me feature is disabled at the component level.
 * HTML5 number field for 6 digit codes (Email, PushBullet, SMS, Time-based One Time Password)
+* Now using WebAuthn library version 2, required for operating inside Joomla 4
 
 **Bug fixes**